+++ eslint.config.mjs
@@ -24,7 +24,9 @@ export default tseslint.config(
     ignores: [
       "node_modules",
       "dist",
+      "public",
       "coverage",
+      "public",
       "vite.config.ts",
       "eslint.config.mjs",
       "avclient-livekit.js",